libwatchfish notifications
Switched notification monitoring to libwatchfish NotificationMonitor,
which can track the notifications better, detect when they're closed and
close them. The libpebble Notification now creates its UUID itself so
that the platform can keep hold of it. Selecting 'Dismiss' on the watch
will close the notification on the Sailfish events view. 
A signal is emitted when the phone removes a notification, but nothing
is done with it on the libpebble side yet. It should remove the
notification there. The parent Id of a TimelineItem is now settable, and
is set to a special value for notifications, else the dismissal is not
reported to the phone.