Music PlayState support
Add missing fields to MusicControlUpdateCurrentTrack and support for the
second music packet - MusicControlUpdatePlayStateInfo, which contains
trackPosition and status info so the watch can display a progress bar.
Harmonize the MusicControl enum with libpebble2.
On Sailfish, position seeking does not seem to be reported from
MprisManager, so the watch will be out of sync when this happens.