Notification changes
Store filters by owner instead of name to avoid locale issues, but also
store the app name for display. Drop the separate icon ini files and
change the notifications.ini format to store a section per app, so we
can store the name and the icon.
If the icon or name are missing in the notification, search for the
source in the launchers rather than doing it from the UI.
Remove hard-coded icons, we shouldn't need those now.
Emit notifications when they're changed as well as when they're new.
Don't bother setting the sender if the icon gives the same info.
Bit of a hack: drop any notification with a category ending in '.group'
as we will have also been sent the real message alongside. Still need to
find out how the lipstick events view filters these...